Refactor: Id 타입을 Long으로 변경
Integer->Long으로 변경한 이유: int, long이 아닌 Wrapper Class(Long)를 쓰는 이유는 null을
사용할 수 있기 때문. primitive type은 기본값이 0이 되는데, id가 없을 때를 나타낼 수 없음(long은
primitive type으로, 값이 없을 경우 0으로 초기화) 즉, null 을 표현할 수가 없음. 또한 Long과
Integer 중에서 Long을 쓰는 이유는 표현범위 때문임.

+++ src/main/java/kr/app/restfulapi/uga/post/controller/PostController.java
@@ -44,7 +44,7 @@ public class PostController {
   }
 
   @GetMapping("/{id}")
-  public ResponseEntity<EntityModel<PostDto>> getPostById(@PathVariable Integer id,
+  public ResponseEntity<EntityModel<PostDto>> getPostById(@PathVariable Long id,
       @AuthenticationPrincipal UserDetails userDetails) {
 
     Optional<PostDto> optionalPostDto = postService.getPostById(id, userDetails);
@@ -70,7 +70,7 @@ public class PostController {
   }
 
   @PutMapping("/{id}")
-  public ResponseEntity<PostDto> updatePost(@PathVariable Integer id,
+  public ResponseEntity<PostDto> updatePost(@PathVariable Long id,
       @RequestBody @Valid PostDto postDto, @AuthenticationPrincipal UserDetails userDetails) {
 
     Optional<PostDto> updatedPost = postService.updatePost(id, postDto, userDetails);
@@ -79,7 +79,7 @@ public class PostController {
   }
 
   @DeleteMapping("/{id}")
-  public ResponseEntity<Void> deletePost(@PathVariable Integer id,
+  public ResponseEntity<Void> deletePost(@PathVariable Long id,
       @AuthenticationPrincipal UserDetails userDetails) {
 
     boolean deleted = postService.deletePost(id, userDetails);
