+++ src/main/java/kr/app/restfulapi/uga/post/dto/PostDto.java
@@ -3,7 +3,7 @@ package kr.app.restfulapi.uga.post.dto;
 import kr.app.restfulapi.uga.post.entity.Post;
 import kr.app.restfulapi.uga.user.entity.User;
 
-public record PostDto(Integer id, String title, String content) {
+public record PostDto(Long id, String title, String content) {
 
   static public PostDto toDto(Post post) {
     return new PostDto(post.getId(), post.getTitle(), post.getContent());
